+++ backend/marker/src/report.rs
@@ -70,9 +70,9 @@ use serde::Serialize;
 #[derive(Debug, Serialize, Clone)]
 pub struct Score {
     /// Points earned by the student.
-    pub earned: i64,
+    pub earned: f32,
     /// Total possible points.
-    pub total: i64,
+    pub total: f32,
 }
 
 /// Represents a subsection of a grading task, such as a subtask or rubric item.
@@ -81,9 +81,9 @@ pub struct ReportSubsection {
     /// Label or name of the subsection (e.g., "Subtask 1").
     pub label: String,
     /// Points earned for this subsection.
-    pub earned: i64,
+    pub earned: f32,
     /// Total possible points for this subsection.
-    pub total: i64,
+    pub total: f32,
     /// Feedback or comments for this subsection.
     pub feedback: String,
 }
@@ -116,9 +116,9 @@ pub struct CoverageFile {
     /// File path (relative or absolute).
     pub path: String,
     /// Lines or items covered in this file.
-    pub earned: i64,
+    pub earned: f32,
     /// Total lines or items in this file.
-    pub total: i64,
+    pub total: f32,
 }
 
 /// The top-level grading report, aggregating all tasks, scores, and optional code analysis results.
@@ -191,16 +191,16 @@ mod tests {
 
     fn sample_score() -> Score {
         Score {
-            earned: 8,
-            total: 10,
+            earned: 8.0,
+            total: 10.0,
         }
     }
 
     fn sample_subsection() -> ReportSubsection {
         ReportSubsection {
             label: "Subtask 1".to_string(),
-            earned: 4,
-            total: 5,
+            earned: 4.0,
+            total: 5.0,
             feedback: "Good job".to_string(),
         }
     }
@@ -217,11 +217,11 @@ mod tests {
     #[test]
     fn test_score_struct_fields() {
         let score = Score {
-            earned: 5,
-            total: 10,
+            earned: 5.0,
+            total: 10.0,
         };
-        assert_eq!(score.earned, 5);
-        assert_eq!(score.total, 10);
+        assert_eq!(score.earned, 5.0);
+        assert_eq!(score.total, 10.0);
     }
 
     #[test]
@@ -231,14 +231,14 @@ mod tests {
             task_number: 2,
             name: "Task 2".to_string(),
             score: Score {
-                earned: 7,
-                total: 10,
+                earned: 7.0,
+                total: 10.0,
             },
             subsections: vec![subsection.clone()],
         };
         assert_eq!(task.task_number, 2);
         assert_eq!(task.name, "Task 2");
-        assert_eq!(task.score.earned, 7);
+        assert_eq!(task.score.earned, 7.0);
         assert_eq!(task.subsections[0].label, "Subtask 1");
     }
 
@@ -276,15 +276,15 @@ mod tests {
     fn test_generate_new_mark_report_function() {
         let now = "2024-06-01T12:00:00Z".to_string();
         let mark = Score {
-            earned: 10,
-            total: 20,
+            earned: 10.0,
+            total: 20.0,
         };
         let tasks = vec![sample_task()];
         let report =
             generate_new_mark_report(now.clone(), now.clone(), tasks.clone(), mark.clone());
         assert_eq!(report.created_at, now);
         assert_eq!(report.updated_at, now);
-        assert_eq!(report.mark.earned, 10);
+        assert_eq!(report.mark.earned, 10.0);
         assert_eq!(report.tasks.len(), 1);
     }
 
@@ -292,13 +292,13 @@ mod tests {
     fn test_code_coverage_optional_fields() {
         let coverage = CodeCoverageReport {
             summary: Some(Score {
-                earned: 5,
-                total: 10,
+                earned: 5.0,
+                total: 10.0,
             }),
             files: vec![CoverageFile {
                 path: "src/lib.rs".to_string(),
-                earned: 5,
-                total: 10,
+                earned: 5.0,
+                total: 10.0,
             }],
         };
         let report = MarkReport {
